refactor: reorganize folder structure in @/components
- Refactored the directory structure within `@/components` for better organization and clarity.
- Grouped related components into dedicated folders to enhance maintainability and scalability.
- Updated imports across the codebase to reflect the new folder structure.
