+++ components/header/menu/user/user-menu-item-sign-out.tsx
