+++ components/header/shop-header.tsx
@@ -1,0 +1,10 @@
+import { MainHeader } from "@/components/header/main-header";
+import { HeaderMenu } from "@/components/header/menu/header-menu";
+
+export const ShopHeader = () => {
+  return (
+    <MainHeader>
+      <HeaderMenu />
+    </MainHeader>
+  );
+};
