--- components/layout/header/shop-header.tsx
@@ -1,10 +1,0 @@
-import { MainHeader } from "@/components/layout/header/main-header";
-import { HeaderMenu } from "@/components/layout/header/menu/header-menu";
-
-export const ShopHeader = () => {
-  return (
-    <MainHeader>
-      <HeaderMenu />
-    </MainHeader>
-  );
-};
